[fix] Rename FamiliesPage to FamilyPage
