+++ client/src/pages/Family/FamilyPage.test.tsx
@@ -2,16 +2,16 @@ import { fireEvent, render, screen, waitFor } from "@testing-library/react";
 import React from "react";
 import { act } from "react-dom/test-utils";
 
-import FamiliesPage from "./FamiliesPages";
+import FamilyPage from "./FamilyPage";
 
 test("renders Family page", () => {
-  render(<FamiliesPage />);
+  render(<FamilyPage />);
   const textElement = screen.getByText("Family");
   expect(textElement).toBeInTheDocument();
 });
 
 test("renders confirmation when clicking delete", async () => {
-  render(<FamiliesPage />);
+  render(<FamilyPage />);
   const removeMateBtn = screen.getAllByLabelText("delete")[0];
   await act(async () => {
     fireEvent.click(removeMateBtn);
